# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Leasing Operations Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /internal/process-interaction/{id}:
    post:
      tags: [Internal]
      operationId: processInteraction
      summary: Trigger AI extraction for an interaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Processing started
...